Create GuessingGame.java
Creating a Guessing Game

Building a program that generates a random number and challenges the user to guess it. The program prompt the user to input their guess, compare it to the generated number, and provide feedback if the guess is too high or too low. It continues until the user correctly guesses the number and then it is displaying the number of attempts it took to win the game.